Add table of contents to docs
This change causes the docs to be split into three parts. The first
part is reserved for the header and title information. The second
is the table of contents themselves, and the last is the content. This
does not include the dependencies list because that information is not
relevant except in the README.md

The indices used in the view are there
to find the spliting points. This is not an optimal solution, but it
will work.