Generallize core construction
This generalization makes it easy to make new cores based off their
name rather than a predefined value.

This new method splits the Cores into a tuple of (Type, X, Y) values.
The Type can be one of T or C, the X can, in theory, be any atom with
that allows at least 2 bonds. The Y value can be any atom that allows
at least 3 bonds.

With this commit is the addition of the script to convert mol2 files
into usable data files.