Add auto selection of proper command base for clusters
The bases are needed because for some reason all the qcommands for
running/viewing/deleting jobs on the clusters are defined in shell
scripts. This specific fix is needed to remove the errors from using
all the bases like it was before. This will help with error handling
down the road.