Remove all functions/views dealing with reseting jobs
This includes removing the functions to pull compressed files from
the cluster. The reason for this is, those functions are not really
ideal. Overall, resetting jobs is not going be supported for now.

Instead of using those functions, it might be better to use a rsync
and a cron job.