Update dataparser so that it is a little easier to use with an API
This change moves the curve fitting into a separate function so it
can be used in other parts of the code.

In addition, there is some simplification in writing the output.
This change adds an extra line before the error messages. This is
why there is the added line in output.txt