Fix warning on where render_to_string does not include csrf_token
This gets rid of the following warning when you try to GET
/chem/jobs/running.json?html=true

UserWarning: A {% csrf_token %} was used in a template, but the context
did not provide the value.  This is usually caused by not using
RequestContext. warnings.warn("A {% csrf_token %} was used in a
template, but the context did not provide the value.  This is usually
caused by not using RequestContext.")