Add created time to FeatureVector
This is done so that it is easy to select when using load_data. By
doing this, all of the newly created feature vecctors can be selected
at the same so long as they are all given the same created time.
Note: This also uses the same hack to get around the auto_now_add
as seen in f8b26b9622e7.