Update load_data to do load M2M connection between feat vecs and data
There are many subtle things in this commit, so I will try to enumerate
all of them.

1) now = timezone.now()
This is used as a way to basically guarantee that all of the new
datapoints and feature vectors will be selected together.

2) preexist = set(FeatureVector...)
This is used because to trim down the number of times the DB has to
be accessed when creating new feature vectors.

3) *.clean_fields()
These calls are needed to make sure that the model is valid before
passing it off to the bulk_create methods. This is because they will
blow up if there are any errors.

4) Through = DataPoint.vectors.through
This makes use of Django's through models to directly bulk_create
all of the M2M connections between the datapoints and the feature
vectors.

5) if len(x) > y:
These are all because of an issue with sqlite having more than
999(?) parameters per database call. To work around this, these
checks are used to limit the size of each call.