Add automatic freezing of dihedral angles between rings if rotated
The angles are only frozen if an angle is given. This allows the
creation of flip structures without them being frozen, and it allows
creating flips that are frozen with (0) and (180) respectively.

This addition also includes the addition of a new method on the Atom
class that gets all the atoms that the atom is connected to.