Finally fix all the issues with windows files
This is a proper fix for working with Windows files. Instead
of determining the file type at the parser level, this is
determined by the log based on the inclusion of '\r' characters.

This is then read from the main Log object using a dynamic
property on the parser which looks at the state given by the log.
This comes with changes to the tests because there are differences
in precision based on where the energies are coming from. For now,
this will continue to stay fairly ambiguous.