Add conversion from Standard to Input form
This is an annoying change to account for times when Gaussian does not print
out an input form geometry at every step. It ends up that it can be a
mixture of both (I+S, I, or S). This is done by doing Procrustes analysis
on the two different geometries to determine the linear transformation
between the two.