Move to using Mocks for cluster tests
This is a huge change that has been needed for a long time. All the
previous calls to an external server in the cluster tests have been
replaced with mock calls. This change needs some refinement, but is a
good staging point for future refinements.

There is still the potential for adjusting these tests such that they
allow for full integration style tests at some point, but for now this
is fine.