Add ability to look at percent fo orbitals on different axes
This feature was built a long time ago as part of the original chemtools
set of utilities. At the time, it was a very naive implementation that
only sort of worked. This new implementation is more correct, but it
makes the assumption that the molecule is a benzobisazole.

In high level terms, the algorithm first identifies the core of the
benzobisazole by finding all the ring structures in the molecule, and
then it filters them based on if they match the core pattern. Then, it
uses a breadth first search to find all the nodes extending from the
top, bottom, left, and right of the molecule. These are then used to
select the correct orbital indices for computation.